Update __init__.py
Decrease the number of rounds for bcrypt to speed up checking of user codes.

This decreases the amount of effort an attacker would need to brute force user codes. Given that an attacker would need physical access to a HA instance or an encrypted backup, this seems like an acceptable tradeoff.